sam-todaylaw/TL_bookmark_add/requirements.txt
requests
boto3
pymongo
PyJWT